fusion con develop-ian-copia

+++ cfp-mapa/src/main/java/com/cfp/mapa/dto/metricas/RendimientoDTO.java
@@ -7,7 +7,7 @@ public record RendimientoDTO(
     TipoReporte tipo,
     long creados,
     long cerrados,
-    double promedioMinutos,
+    long promedioMinutos,
     long maximoMinutos
 
 ) {
